TS-1150 Limit headers to 64KB, which is our internal limit anyways

+++ proxy/hdrs/MIME.h
@@ -678,7 +678,7 @@ int mime_mem_print(const char *src_d, int src_l, char *buf_start, int buf_length
 int mime_field_print(MIMEField * field, char *buf_start, int buf_length,
                      int *buf_index_inout, int *buf_chars_to_skip_inout);
 
-const char *mime_str_u16_set(HdrHeap * heap, const char *s_str, uint16_t s_len,
+const char *mime_str_u16_set(HdrHeap * heap, const char *s_str, int s_len,
                              const char **d_str, uint16_t * d_len, bool must_copy);
 
 int mime_field_length_get(MIMEField * field);
